refactor(solver): split interner thread-local cache
